requirements.txt
argparse
pulumi>=3.0.0,<4.0.0
pulumi-aws>=4.0.0,<5.0.0
git+https://github.com/kjenney/encryptor.git
pyaml>=21.8.3


